refactor(templates): drop the v2 suffix from the invoice package
With the classic invoice stack gone, the layered presets no longer need the
v2 package segment. Move templates.invoice.v2.presets to
templates.invoice.presets and rewrite every import and doc reference across
the engine, tests, examples, and benchmarks. Pure rename — the visual parity
test passes against the unchanged committed baseline, so the render is
identical.

Tests: ./mvnw verify javadoc:javadoc -pl . — 1378 tests, 0 failures, javadoc
clean; examples + benchmarks compile; perf-smoke + examples-generation smoke
(85) green.
