+++ benchmarks/src/main/java/com/demcha/compose/AllocationRateProbe.java
@@ -5,7 +5,7 @@ import com.demcha.compose.document.api.DocumentSession;
 import com.demcha.compose.document.templates.api.DocumentTemplate;
 import com.demcha.compose.document.templates.data.invoice.InvoiceDocumentSpec;
 import com.demcha.compose.document.templates.data.proposal.ProposalDocumentSpec;
-import com.demcha.compose.document.templates.invoice.v2.presets.ModernInvoice;
+import com.demcha.compose.document.templates.invoice.presets.ModernInvoice;
 import com.demcha.compose.document.templates.proposal.v2.presets.ModernProposal;
 
 import java.lang.management.GarbageCollectorMXBean;
