+++ examples/src/main/java/com/demcha/examples/templates/invoice/InvoiceCinematicFileExample.java
@@ -6,7 +6,7 @@ import com.demcha.compose.document.api.DocumentSession;
 import com.demcha.compose.document.templates.api.DocumentTemplate;
 import com.demcha.compose.document.templates.core.theme.BrandTheme;
 import com.demcha.compose.document.templates.data.invoice.InvoiceDocumentSpec;
-import com.demcha.compose.document.templates.invoice.v2.presets.ModernInvoice;
+import com.demcha.compose.document.templates.invoice.presets.ModernInvoice;
 import com.demcha.examples.support.ExampleDataFactory;
 import com.demcha.examples.support.ExampleOutputPaths;
 
@@ -14,7 +14,7 @@ import java.nio.file.Path;
 
 /**
  * Runnable showcase for the cinematic invoice look on the layered
- * {@code invoice.v2} surface — {@link ModernInvoice} on
+ * layered invoice surface — {@link ModernInvoice} on
  * {@link BrandTheme#invoiceModern()}, rendering the shared
  * {@link InvoiceDocumentSpec} sample on the cream page background.
  *
