+++ src/main/java/com/demcha/compose/document/templates/data/invoice/package-info.java
@@ -1,5 +1,5 @@
 /**
  * Shared, render-neutral invoice document specs and supporting data
- * records, consumed by the layered {@code invoice.v2} presets.
+ * records, consumed by the layered {@code invoice.presets} presets.
  */
 package com.demcha.compose.document.templates.data.invoice;
