+++ src/main/java/com/demcha/compose/document/templates/invoice/presets/ModernInvoice.java
@@ -1,4 +1,4 @@
-package com.demcha.compose.document.templates.invoice.v2.presets;
+package com.demcha.compose.document.templates.invoice.presets;
 
 import com.demcha.compose.document.api.DocumentSession;
 import com.demcha.compose.document.dsl.TableBuilder;
