+++ src/main/java/com/demcha/compose/document/templates/invoice/presets/package-info.java
@@ -14,4 +14,4 @@
  *
  * @since 2.0.0
  */
-package com.demcha.compose.document.templates.invoice.v2.presets;
+package com.demcha.compose.document.templates.invoice.presets;
